fix organizer edit event crashing and moved edit and view events to activity folder
