+++ Soroban/app/src/main/java/com/example/soroban/activity/OrganizerDashboardActivity.java
@@ -3,21 +3,17 @@ package com.example.soroban.activity;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.ListView;
 import android.widget.TextView;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.soroban.OrganizerEventViewDetailsActivity;
 import com.example.soroban.R;
 import com.example.soroban.adapter.EventArrayAdapter;
 import com.example.soroban.model.Event;
 import com.example.soroban.model.EventList;
 import com.example.soroban.model.User;
-
-import java.util.Calendar;
 
 public class OrganizerDashboardActivity extends AppCompatActivity {
