+++ Soroban/app/src/main/java/com/example/soroban/activity/OrganizerEventEditDetailsActivity.java
@@ -1,17 +1,15 @@
-package com.example.soroban;
+package com.example.soroban.activity;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.google.android.material.textfield.TextInputEditText;
+import com.example.soroban.R;
 
 public class OrganizerEventEditDetailsActivity extends AppCompatActivity {
     private EditText eventTitle;
