+++ Soroban/app/src/main/java/com/example/soroban/activity/OrganizerEventViewDetailsActivity.java
@@ -1,4 +1,4 @@
-package com.example.soroban;
+package com.example.soroban.activity;
 
 import android.content.Intent;
 import android.graphics.Bitmap;
@@ -18,14 +18,14 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
 
-import com.example.soroban.activity.EventEntrantsListActivity;
+import com.example.soroban.FireBaseController;
+import com.example.soroban.QRCodeGenerator;
+import com.example.soroban.R;
 import com.example.soroban.model.Event;
 import com.example.soroban.model.User;
 
 import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.Locale;
-import java.util.Objects;
 
 public class OrganizerEventViewDetailsActivity extends AppCompatActivity {
     private Event selectedEvent;
